Changed the home screen
I changed the home screen on the app. I found from test with other apps,
using the hamburger menu as the main menu isn't a good idea for such a
simple app. Most users don't know to click on it, and it makes it harder
for the average user.
So I'm making the main menu into a simple button menu